+++ src/main/java/net/austizz/ultimate_auction_system/api/remote/UasRemoteCommandStatus.java
@@ -1,0 +1,6 @@
+package net.austizz.ultimate_auction_system.api.remote;
+
+public enum UasRemoteCommandStatus {
+    SUCCEEDED,
+    FAILED
+}