+++ src/main/java/net/austizz/ultimate_auction_system/api/remote/UasRemoteCommandType.java
@@ -1,0 +1,7 @@
+package net.austizz.ultimate_auction_system.api.remote;
+
+public enum UasRemoteCommandType {
+    BID,
+    BUYOUT,
+    CANCEL
+}